pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-eval-cli"
version = "0.1.0"
description = "A lightweight AI workflow evaluator."
requires-python = ">=3.9"
dependencies = [
    "google-genai",
    "pydantic",
    "loguru",
    "rich"
]

[project.scripts]
ai-eval = "ai_eval_cli.cli.runner:main"